search class

+++ private/controllers/Schools.php
@@ -10,14 +10,18 @@ class Schools extends controller
             $this->redirect('login');
         }
         $school = new school();
-       
-        $data = $school->findAll();
 
-        $crumbs[] = ['Dashboard', ''];
-        $crumbs[] = ['schools', 'schools'];
-          $this->view('schools',[
-            'crumbs'=>$crumbs,
-            'rows'=>$data]);
+            $data = $school->findAll();
+
+            $crumbs[] = ['Dashboard', ''];
+            $crumbs[] = ['schools', 'schools'];
+         if(Auth::access('superAdmin')){
+                $this->view('schools',[
+                    'crumbs'=>$crumbs,
+                    'rows'=>$data]);
+        }else{
+            $this->view('access_denied');
+        }
     }
     
     //add school
@@ -29,7 +33,7 @@ class Schools extends controller
 
         }
         $errors = array();
-        if(count($_POST) > 0)
+        if(count($_POST) > 0 && Auth::access('superAdmin'))
         {
         $school = new school();
         if($school->validate($_POST)){
@@ -49,10 +53,15 @@ class Schools extends controller
              $crumbs[] = ['Dashboard', ''];
              $crumbs[] = ['schools', 'schools'];
              $crumbs[] = ['add', 'schools/add'];
+       
+      if(Auth::access('superAdmin')){
           $this->view('schools.add',[
             'errors'=>$errors,
             'crumbs'=>$crumbs,
             ]);
+        }else{
+            $this->view('access-denied');
+        }
     }
 
 //edit school
@@ -64,8 +73,9 @@ class Schools extends controller
         }
 
         $school = new school();
+
         $errors = array();
-        if(count($_POST) > 0)
+        if(count($_POST) > 0 && Auth::access('superAdmin'))
         {
         
         if($school->validate($_POST)){
@@ -87,11 +97,16 @@ class Schools extends controller
              $crumbs[] = ['Dashboard', ''];
              $crumbs[] = ['schools', 'schools'];
              $crumbs[] = ['edit', 'schools/edit'];
-          $this->view('schools.edit',[
-            'row'=>$row,
-            'errors'=>$errors,
-            'crumbs'=>$crumbs,
-            ]);
+
+         if(Auth::access('superAdmin')){
+            $this->view('schools.edit',[
+                'row'=>$row,
+                'errors'=>$errors,
+                'crumbs'=>$crumbs,
+                ]);
+        }else{
+            $this->view('access_denied');
+        }
     }
 
 
@@ -106,7 +121,7 @@ class Schools extends controller
         $school = new school();
 
         $errors = array();
-        if(count($_POST) > 0)
+        if(count($_POST) > 0 && Auth::access('superAdmin'))
         {
             $school->delete($id);
             $this->redirect('schools');
@@ -118,10 +133,15 @@ class Schools extends controller
             $crumbs[] = ['Dashboard', ''];
             $crumbs[] = ['schools', 'schools'];
             $crumbs[] = ['delete', 'schools/delete'];
+     
+        if(Auth::access('superAdmin')){
           $this->view('schools.delete',[
             'row'=>$row,
             'crumbs'=>$crumbs,
             ]);
+        }else{
+            $this->view('access_denied');
+        }
     }
 }
